+++ CHANGELOG.md
@@ -5,10 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
-## [8.15.0]
-
 ### Added
-- Adding support for load balancer redirect rule status
+- managed load balancer: support for redirect rule HTTP status
 
 ## [8.14.0]