+++ CHANGELOG.md
@@ -7,6 +7,7 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ### Added
 - Added support for token-based authentication in the client, including functions for token management.
+- managed load balancer: support for redirect rule HTTP status
 
 ## [8.14.0]